Logging fix, thread fix, added created timestamp, prime number bandaid
Lots of fixes. May require cleanup. fixed logging stacking infinitely. Reduced the payload in the query to get players. Made it so it will grab only the players in a thread if called from a thread. Added created on timestamp to the bottom. Runs to get players twice, divided by 2 prime numbers to get as many users as possible.